+++ my_array.rb
@@ -1,0 +1,5 @@
+require_relative 'my_enumerable'
+
+class MyArray < Array
+  include MyEnumerable
+end
